Solved issues with compiling python bindings on stampede2 (and running the test suite).
Previously, the python3 installation would not be able to find the pygrackle module after installation. The cython language level had not been specified and by default it assumed a language_level of 2. I found that specifying a language level of 3 fixed the problem.

To that end, I have added a few lines to setup.py to set the cython language to match the major version number of the python installation. I only needed to modify a single line in grackle_wrapper.pyx to support language_level=3; I changed an implicit relative import to an explicit relative import (see https://stackoverflow.com/a/54901312 for details).