Introduced test to explicitly check that members of the chemistry_data struct are synchronized with the parameters used by the dynamic API.
This test makes use of a program called castxml (https://github.com/CastXML/CastXML). This is a program (built on top of LLVM/Clang) that parses C++ code and outputs the syntax tree of all of all C++ structs/classes (their fields and methods), namespaces, function declarations, and global variables.

The new test executes castxml on the header containing the declaration of the chemistry_data struct and parses the output xml to identify all of the members of chemistry_data (along with their associated types). Then it confirms that there is a corresponding parameter offered by the dynamic api that can be used to configure that member.

The main idea is for this test to be run on circleci (I don't expect users to actually install castxml). Consequently, the test will be skipped with an explanation if the castxml binary can't be found.

Note that there is a python wrapper for castxml that is developed by its creators, called pygccxml (as an aside, castxml's predecessor was called gccxml - it was built on top of gcc instead of LLVM/clang). I opted not to require that as a dependency (at least initially) to avoid potential python module compatability issues that could crop up down the road, but it may simplify the testing code. But, I'm happy to change this