Introduce the grackle_num_params function to let you query the number of parameters of a given type that are stored in chemistry_data.
When I first implemented the dynamic-api, I omitted this function because I didn't anticipate it being particularly useful and I wanted to avoid expanding the API by too much.

  - Indeed, I didn't really feel like I needed this function while refactoring Grackle's python bindings.

  - However, it turns out that this functionallity is VERY useful when using the dynamic-api to write code in a statically typed language like C/C++ (while trying to minimize the definition of global variables). It's especially useful to know the number of parameters of a given type, before iterating over the parameters of that type during serialization.

The new function has the following signature:

```
size_t grackle_num_params(const char* type_name);
```

This function returns the number of parameters of a given type when it is passed ``"int"``, ``"double"``, or ``"string"``. For any other argument, this returns ``0``. I opted NOT to define a separate function for each datatype to avoid introducing more functions than necessary to Grackle's public API (and the added cost of the string comparison is minimal).

While I was introducing this new functionallity, I refactored the rest of the dynamic API. All of the dynamic api functions are faster than they were previously. Notably, the complexity of the ``param_name_*`` functions was reduced from ``O(N)`` (where ``N`` is the total number of parameters) to ``O(1)``. As a point of reference, I had previously discussed this alternative implementation back when we first introduced the Dynamic API (PR #130). I made the change now because the older implementation approach did not lend itself to implementing ``grackle_num_params``.