Introduced the GRACKLE_USE_OMP macro to grackle_config.h
This macro is always assigned a value of 0 or 1, which denotes whether Grackle is being used. This macro is used to automatically control whether the omp_nthreads field is defined in external programs. External programs can also directly query the macro's value at compile-time to determine whether Grackle was compiled with openmp