Altered calc_temp1d_cloudy_g's signature to avoid logical type
This was done to circumvent some issues with transcribing that subroutine to C. Specifically, the LOGICAL dtype is incompatible with C, unless we explicitly use LOGICAL(KIND=C_BOOL) everywhere. Since this may also produce issues with C++ (the bool type in C++ and the _Bool type introduced in C99 may not be compatible), we have opted to replace it with an integer type.

In practice, the only change we needed to make to calc_temp1d_cloudy_g's signature was to alter the type of itmask. To make this all work, without changing large regions of the codebase, we have introduced a wrapper function called calc_temp1d_cloudy_shim_g.F that explicitly converts the dtype from LOGICAL to INTEGER of itmask. This introduces an extra allocation, but I'm skeptical it will make a huge difference in speed (at least right now - as the code is currently written)