Make public types refer to named structs
Prior to this PR, all types that alias structs alias anonymous structs.
This is totally fine, but it prevents users from forward declaring
types.

In this PR, we make:
- `chemistry_data` an alias for `struct chemistry_data`
- `code_units` an alias for `struct code_units`
- `chemistry_data_storage` an alias for `struct chemistry_data_storage`
- `grackle_field_data` an alias for `struct grackle_field_data`

By doing this, downstream codes can now forward declare types like
`struct chemistry_data`, regardless of whether grackle is being linked
in the current build, which is useful for reducing ifdef statements

In Enzo-E, I'm looking to use this feature so that we don't have to
rebuild the entire program if we decide to rebuild with/without linking
against Grackle