Add error-check on `grackle_field_data.grid_dx`
The `grid_dx` field is rarely needed for grackle operations and it sometimes doesn't make sense for a downstream application to properly set the value (e.g. where cells are not perfect cubes or if the code isn't grid-based).

As part of this commit, I actually factored the error-check out into it's own function (as well as another related check) in order to make sure that we apply the same test in both the `calculate_cooling_time` and `solve_chemistry` functions.