Introduce hdf5impl of visitor-state-dump-logic
Previously, we were dumping `chemistry_data`, `code_units`, and `grackle_version` to a json-string that was stored in the hdf5 file. While this certainly got the job done, there were 2 drawbacks:

1. we did it in a rather round-about manner...
2. if anyone ever wanted to deserialize the dumps into new instances, you would need to support hdf5 reading AND json-parsing

With this new change, we now dump these structs in a way that individual key-value pairs are stored as attributes. This simplifies hypothetical deserialization