close gap between build-systems
This commit adds capabilities to perform a basic installation with CMake. It also adopts a policy consistent with the older build-system for naming shared libs (and making symbolic links).

This policy is a little unidiomatic, but it is probably a better strategy than what idiomatic CMake could produce (given that we don't support ABI compatability)