Remove `tail` invocation from `config/query_version.py`
The required logic to fetch the last line of the file is now implemented in python. The logic loops over all lines from the VERSION file (from beginning to end) and returns the final non-empty line. This is fine for the VERSION file which should really just be 1 line (but is suboptimal for large files)