Introduce GR_SUCCESS and GR_FAIL
These are macros in the public grackle.h header file. The idea is that
external applications will be able to decode the error-codes in a
self-consistent manner with the rest of Grackle.

In practice, the main benefit here is that the python bindings no longer
need to include a private Grackle header. This will ultimately lead to a
minor simplification in GH #208.

Ideally, we would go through and replace all occurences of ``SUCCESS``
with ``GR_SUCCESS`` and ``FAIL`` with ``GR_FAIL`` (so that we can
entirely remove ``SUCCESS`` and ``FAIL`` from the codebase), but that
would be a fairly invasive change that would touch a lot of code.

- I illustrated what that would look like in
  ``initialize_chemistry_data.c`` and I also added a compile-time check
  to ensure consistency of ``GR_SUCCESS`` with ``SUCCESS`` and
  ``GR_FAIL`` with ``FAIL``.

- I'm happy to propoagate this change, but I want some buy-in before I
  pursue that. (Maybe it would be better in a separate PR?)