Remove the lone `cimport numpy` line
This should let us drop numpy as a build-time requirement. This involved
a minor edit to the ``get_field`` c-function written in
grackle_wrapper.pyx (we now use Cython's built-in typed memoryview
functionality instead. While I was there, I made a minor tweak to the
function so that it will perform proper error-checking for a plausible
failure-mode