Remove where clause from itmask_metal calc
In more detail, this is the only where clause in the entire codebase and
for that reason, transcription to C/C++ is easier if we just don't have
it at all.

While I was here, I also replace `itmask_metal = .false.` with an
explicit do-loop because (again) it will simplify C/C++ transcription.
- Yes, I know that there are a number of parts of this branch that use
  array based operations, but I think this excludes most of the contents
  of cool1d_multi_g