Porting over various physical constants defined in the Fortran
A lot of the physical constants in the C and Fortran layers are defined
in slightly different ways. This commit seeks to make all of the Fortran
constants in one form or another in a way that won't produce
name-collisions with the existing constants defined in C.

The biggest difference for most constants is the precision of the
constant.

In the future, I would like to reduce the number of variants of
constants, but that will be easier after we finish the conversion.