change `L{}_(N|Size)` from `int` -> `long long`
This was done to remove a bunch of `int(<arg>, 8)` casts. These removed
casts all occured while passing arguments to `interpolate_Nd_g`
routines. Essentially, the Fortran compiler understands that we want to
cast an argument and then pass the result into the subroutine by
reference. This is tricky to transcribe to C/C++.