Reorganize `L{s}_{prop}` into structs
Prior to this commit, the `chemistry_data_storage` struct declared a 9
attributes for each species ``s``, in the list: ``H2``, ``HD``, ``CI``,
``CII``, ``OI``, ``CO``, ``OH``, ``H2O``. In particular the declaration
looked like:

```c++
{
  long long *Ls_N, Ls_Size;
  double *Ls_D, *Ls_T, *Ls_H, Ls_dD, Ls_dT, Ls_dH, *Ls_L;
}
```

This commit replaced the set of 9 attributes for each species with a
struct called ``gr_interp_grid``. This provides a few benefits:
1. we have removed 64 members from ``chemistry_data_storage``
   (this alone is probably enough reason for the change simplicity is
   better in this complex data structure)
2. we are able to standardize the logic for zero-initializing the data
   in the replaced attributes and for freeing any associated memory
3. this will reduce the code we will eventually need to expose this
   information through the pygrackle interface
4. after transcribing Fortran to C, this will help with encapsulating
   some common operations.

> [!NOTE]
> Technically, most members of ``gr_interp_grid`` are sorted into a
> separate struct called ``gr_interp_grid_props``. The provided comments
> explain that this lays the groundwork for reducing duplication with
> the ``cloudy_data`` struct at some point in the future.

Even if we decide to reorganize this data in the future, it will be easy
to transform it from the new version of the code than it would be before
this commit.