step_rate_newton_raphson: make `ierror` local var
When I previously extracted the `step_rate_newton_raphson` subroutine
out of `solve_rate_cool_g` (back when the latter was still written in
Fortran), I mistakenly thought that `ierror` should be passed as an
argument to `step_rate_newton_raphson` from `solve_rate_cool_g`

It now became apparent to me that `ierror` is ONLY used within
`step_rate_newton_raphson` as a local variable to determine control
flow. As I understand it, `step_rate_newton_raphson` has a while loop
and basically tries to run through the while-loop until it encounters no
errors (and the answer is converged). I don't think
`step_rate_newton_raphson` is able to gracefully fail (that is something
to be addressed in the future).

Consequently, this commit converts `ierror` from an argument of
`step_rate_newton_raphson` to an internal local variable