solve_rate_cool_g: remove comp[12] local variables
Historically, these have passed as arguments to the `cool1d_multi_g`
Fortran subroutine, but it turns out that they are only used internally
as local variable needs to read the value. These variables also needed
to be passed to `step_rate_newton_raphson`, but they are again only
passed to internal calls of `cool1d_multi_g`.

Modifying `step_rate_newton_raphson` and `cool1d_multi_g` is somewhat
beyond the scope of this set of commits, so we currently declare these
variables to the functions that wrap these subroutines with a note
explaining that they really should be local variables within the called
functions