Make GrainSpeciesCollection use a lookup table
This was done to be more consistent with the general-purpose
`SpeciesCollection` type. For concreteness:
- `SpeciesCollection` has always used the enumerators defined within
  `SpLUT` as a compile-time lookup table indices
- this commit refactors `GrainSpeciesCollection` so that it uses the
  the enumerators within the newly-defined `OnlyGrainSpLUT` entity
  as a compile-time lookup table

In the future, this change is going to help simplify the C++ versions
of `calc_all_tdust_gasgr_1d_g` and `calc_grain_size_increment_1d`.
Currently these routines have multiple instances where the same exact
logic is repeated 12 times (once per dust grain species). This change
should help us remove this code-duplication and replace it with a
for-loop.