step_rate_newton_raphson: introduce time_deriv_0d::FrozenSimpleArgs
The idea is to collect args that are forwarded to the time-derivative
calculation and are effectively frozen between various calls

I also converted some of the arguments passed into lookup_cool_rates0d
to read from this pack of args. But, I haven't done it for the really
large structs (e.g. my_chemistry). I think those will be easier to
change later