Relocate the check for whether we are missing the test-answers
Previously, the test_code_examples would eagerly FAIL, if pytest
was invoked to run answer-tests in "compare-mode" and the answer-file
could not be found for the test_code_examples (whether or not we were
planning to skip the test_code_examples).

This commit relocates the FAILURE check until after the logic
determining whether we should skip the test_code_examples.