circleci: formalize distinction between core-lib & pygrackle tests (2/2)
I renamed the "run-tests" command to be called "run-pygrackle-tests". I
also removed the logic for running the OpenMP example from this command

For context, the prior commit configured the "run-core-tests" command
to have the logic for running the OpenMP example. But, to make changes
more atomic, that commit didn't remove the logic from the command now
known "run-pygrackle-tests"