circleci: formalize distinction between core-lib & pygrackle tests (1/2)
Previously, the logic invoking core-lib tests (i.e. through `ctest`) was
implement as a bunch of shell commands directly embedded into a
"run-step" of the test-suite job.

This commit relocates the logic to a named command called
"run-core-tests." This command contains logic for running the
cxx_omp_example example (while this commit simply duplicates the logic,
the next commit will rectify the duplication)