Further loosen up the tolerance
I realized that the running the code on MacOS machine with
CMAKE_BUILD_TYPE=Release and CMAKE_BUILD_TYPE=Debug produces slightly
different answers.

The flags used in these cases are:

  - Release: -O3
  - Debug: -g

This is fairly disturbing since it should not make a difference (-O3
does not enable any floating point optimizations). While I haven't
tested it, I'm **extremely** confident that this isn't a CMake issue.

But I'm not really sure what we can do... Let's keep track of this