make a minor tweak to suppress private-header warnings for unit-tests
Originally I tried to fix the problem properly, by having
grackle_macros.h directly include grackle.h. But, this produced
problems related to the fact that some function are redeclared slightly
differently between grackle.h and within the source files (the
difference is OK since it is just related to the appearance of `const`).
A number of pending PRs gradually fix this disagreement throughtout the
codebase. Thus, we are going to hold off on "fixing it properly" until
those other PRs are merged