Move CastXML parsing to a separate script
This change was made in anticipation of shifting the
"test_grackle_chemistry_field_synched" test from the pygrackle
test-suite to the core-lib test suite.

In more detail, the test doesn't really belong in the pygrackle
test-suite since it is looking at header files. As currently written,
the test current places a strong constraint on the way in which we
organize our public headers (it currently can't read a header that
depends on an autogenerated header file). Shifting this test to the
core-lib test suite will remove this limitation