Mention Enzo-E as an example for auto-install
This patch primarily consists of 2 changes:

1. A modification to the README (and the duplicated text on the landing
   page of the website) mentioning Enzo-E as an example of a code with
   a build system that supports automatically downloading & installing
   Grackle.

2. The addition of a short paragraph at the top of the Installation page
   encouraging people to check the documentation of the simulation code
   that they are using before reading through the entire page. I
   reference the fact that the build systems of some codes (e.g. Enzo-E)
   are able to automatically fetch and install Grackle as part of the
   build process.

I also noticed that the website's landing page had a footnote with
missing text. I added the missing text (it duplicates a footnote in the
README). While I was doing this, I deleted the "Search" section from the
landing page (since it doesn't do anything with the RTD sphinx theme)