Enable Compiler Warnings
This is mostly being done in preparation for transcription in the
newchem-cpp branch.

Background
----------
As we all know, it is very easy to write C/C++ code that is accepted by
the compiler, but is wrong in a subtle manner. For example, C/C++ do
lots of implicit conversions (which can lead to unexpected results).

Goal
----
The goal of this PR is to enable warning flags for the compiler and
treat them as errors. We want to do this in a manner that isn't a burden
to our users. In other words, we want these warnings (and their
treatment as errors) to be off by default and to only turn them on in
CI.

How We Accomplish This
----------------------
To accomplish this, I made a CMakePresets.json file, where we can store
various presets for configuring Grackle-builds. It's designed to give
you lots of customizability for reusing configuration-snippets. This
will be very useful in the future if/when we try configuring sanitizers
(to detect undefined behavior) and possibly for measuring
code-coverage.

As an added bonus, it's also a great starting point for people who
configuring IDEs (like vscode or CLion). People can make their own
personal versions of the presets by creating a CMakeUserPresets.json
file.

**Bottom Line:** this is meant to be a starting point. It would be great
to merge this in ASAP so we can start using this for the C++
transcription PRs.

Documentation
-------------
Let me know if you want me to add narrative docs (I would only add these
to the developer guide. Technically Presets require a newer version of
CMake than our users use. But, I think this is okay for a subset of
workflows run by CI)