Support running testing subset without pygrackle install
Prior to this PR, our test suite would fail unless we had an editable
install. In reality, all test-cases outside of test_models infer the
location of the data files based on the location of the test-files,
which are never part of the wheel.[^1] Therefore, this PR introduces a
few small tweaks to support testing all test-cases outside of
test_models.

Overall, this will be really useful for providing a more rigorous check
of the correctness of our precompiled wheels (see PR #320).

[^1]: I know that some packages do ship their test-suites as part of
      the wheel, but that currently makes absolutely no sense for us
      unless we provide users with an easy way to get the datafiles.
      We can always revisit this choice in the future after we merge
      PRs #235 & #237 (or come up with an equivalent)