CIRCLECI: Consolidate both jobs running pygrackle-test-suite
Originally one job ran the tests after building pygrackle as a
standalone package (i.e. the core Grackle is compiled as part of
pygrackle). The other job ran the tests with versions of pygrackle
where we explicitly built the core library (with both build-systems) as
a separate step from building pygrackle.