Remove last references to Flake8
Background
----------

Previously we had replaced 99% uses of the Flake8 linter with the Ruff
linter (not to be confused with the Ruff formatter). We had continued to
use the Flake8 linter to deal with the following 2 files:

1. **gracklepy/utilities/testing.py**
2. **gracklepy/__init__.py**

In more detail, we were suppressing certain warnings in these 2 files
from Flake8, and transistioning to using the Ruff formatter for these
files would involve making a few changes.

More recently PRs #368 and #369 introduced the required changes into the
main branch. Now that those changes have been ported to the newchem-cpp
branch, we are able to make this change.

What Changed
------------

I essentially changed 3 things:

1. We no longer install and use flake8 in the continuous integration
2. I removed the **setup.cfg** file (it was **only** being used to track
   flake8 configuration information
3. I needed to tweak **pyproject.toml** to both:
   - stop excluding the **gracklepy/utilities/testing.py** and
     **gracklepy/__init__.py** files from the linter
   - stop listing flake8 as a development dependency